requirements.txt
torch
requests
runpod
llava_torch==1.1.1